requirements.txt
torch
torchvision
torchaudio  
opencv-python
pillow
numpy
scikit-learn
matplotlib
flask
flask-cors